fix dashboard-client to match with new data
refactor code && fix pagination
